use new_test folder for universal staker
