+++ new_test/CumulativeFunction.spec.ts
