+++ new_test/helpers/index.ts
@@ -17,7 +17,7 @@ import _ from 'lodash'
 import {
   TestERC20,
   INonfungiblePositionManager,
-  UniswapV3Staker,
+  UniversalV3Staker,
   IUniswapV3Pool,
   TestIncentiveId,
 } from '../../typechain'
@@ -38,7 +38,7 @@ import { TestContext } from '../types'
 export class HelperCommands {
   actors: ActorFixture
   provider: MockProvider
-  staker: UniswapV3Staker
+  staker: UniversalV3Staker
   nft: INonfungiblePositionManager
   router: ISwapRouter
   pool: IUniswapV3Pool
@@ -59,7 +59,7 @@ export class HelperCommands {
     testIncentiveId,
   }: {
     provider: MockProvider
-    staker: UniswapV3Staker
+    staker: UniversalV3Staker
     nft: INonfungiblePositionManager
     router: ISwapRouter
     pool: IUniswapV3Pool
@@ -114,6 +114,7 @@ export class HelperCommands {
       {
         pool: params.poolAddress,
         rewardToken: params.rewardToken.address,
+        rewardCalc: params.rewardCalc,
         ...times,
         refundee: params.refundee || incentiveCreator.address,
       },
@@ -121,7 +122,7 @@ export class HelperCommands {
     )
 
     return {
-      ..._.pick(params, ['poolAddress', 'totalReward', 'rewardToken']),
+      ..._.pick(params, ['poolAddress', 'totalReward', 'rewardToken', 'rewardCalc']),
       ...times,
       refundee: params.refundee || incentiveCreator.address,
     }
@@ -285,6 +286,7 @@ export class HelperCommands {
       await this.staker.connect(incentiveCreator).endIncentive(
         _.assign({}, _.pick(params.createIncentiveResult, ['startTime', 'endTime']), {
           rewardToken: rewardToken.address,
+          rewardCalc: params.createIncentiveResult.poolAddress,
           pool: params.createIncentiveResult.poolAddress,
           refundee: params.createIncentiveResult.refundee,
         })
@@ -315,6 +317,7 @@ export class HelperCommands {
       startTime: params.startTime,
       endTime: params.endTime,
       refundee: params.refundee,
+      rewardCalc: params.rewardCalc,
     })
   }
 
@@ -430,5 +433,6 @@ export const incentiveResultToStakeAdapter: IncentiveAdapterFunc = (params) => (
   startTime: params.startTime,
   endTime: params.endTime,
   rewardToken: params.rewardToken.address,
+  rewardCalc: params.rewardCalc,
   refundee: params.refundee,
 })
