+++ new_test/shared/logging.ts
