+++ new_test/unit/Stakes.spec.ts
@@ -87,6 +87,7 @@ describe('unit/Stakes', () => {
 
       incentiveArgs = {
         rewardToken: context.rewardToken,
+        rewardCalc: context.rewardCalc.address,
         totalReward,
         poolAddress: context.poolObj.address,
         ...timestamps,
@@ -100,6 +101,7 @@ describe('unit/Stakes', () => {
             refundee: incentiveCreator.address,
             pool: context.pool01,
             rewardToken: context.rewardToken.address,
+            rewardCalc: context.rewardCalc.address,
             ...timestamps,
           },
           _tokenId
@@ -235,6 +237,7 @@ describe('unit/Stakes', () => {
               refundee: incentiveCreator.address,
               pool: context.pool01,
               rewardToken: context.rewardToken.address,
+              rewardCalc: context.rewardCalc.address,
               ...timestamps,
             },
             otherTokenId
@@ -251,6 +254,7 @@ describe('unit/Stakes', () => {
               refundee: incentiveCreator.address,
               pool: context.pool01,
               rewardToken: context.rewardToken.address,
+              rewardCalc: context.rewardCalc.address,
               ...timestamps,
               startTime: timestamps.startTime + 10,
             },
@@ -294,6 +298,7 @@ describe('unit/Stakes', () => {
       stakeIncentiveKey = {
         refundee: incentiveCreator.address,
         rewardToken: context.rewardToken.address,
+        rewardCalc: context.rewardCalc.address,
         pool: context.pool01,
         ...timestamps,
       }
@@ -301,6 +306,7 @@ describe('unit/Stakes', () => {
       incentiveId = await helpers.getIncentiveId(
         await helpers.createIncentiveFlow({
           rewardToken: context.rewardToken,
+          rewardCalc: context.rewardCalc.address,
           totalReward,
           poolAddress: context.poolObj.address,
           ...timestamps,
@@ -364,6 +370,7 @@ describe('unit/Stakes', () => {
 
       createIncentiveResult = await helpers.createIncentiveFlow({
         rewardToken: context.rewardToken,
+        rewardCalc: context.rewardCalc.address,
         totalReward,
         poolAddress: context.poolObj.address,
         ...timestamps,
@@ -385,6 +392,7 @@ describe('unit/Stakes', () => {
         {
           refundee: incentiveCreator.address,
           rewardToken: context.rewardToken.address,
+          rewardCalc: context.rewardCalc.address,
           pool: context.pool01,
           ...timestamps,
         },
@@ -484,6 +492,7 @@ describe('unit/Stakes', () => {
 
       createIncentiveResult = await helpers.createIncentiveFlow({
         rewardToken: context.rewardToken,
+        rewardCalc: context.rewardCalc.address,
         totalReward,
         poolAddress: context.poolObj.address,
         ...timestamps,
@@ -520,6 +529,7 @@ describe('unit/Stakes', () => {
         {
           refundee: incentiveCreator.address,
           rewardToken: context.rewardToken.address,
+          rewardCalc: context.rewardCalc.address,
           pool: context.pool01,
           ...timestamps,
         },
@@ -534,6 +544,7 @@ describe('unit/Stakes', () => {
             refundee: incentiveCreator.address,
             pool: context.pool01,
             rewardToken: context.rewardToken.address,
+            rewardCalc: context.rewardCalc.address,
             ...timestamps,
           },
           tokenId
